\name{print.nice}
\alias{print.nice}
\alias{print.nice.default}
\title{Generic print.nice Function and Default Method}

\usage{
print.nice(x, ...)

\method{print.nice}{default}(x, ...)
}

\arguments{
\item{x}{an \R object.}
\item{\dots}{further arguments passed to or from methods.}
}

\description{
Generic print.nice function and default method.
}


