% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.nice.R
\name{print.nice}
\alias{print.nice}
\alias{print.nice.default}
\title{Generic print.nice function}
\usage{
print.nice(x, ...)

\method{print}{nice.default}(x, ...)
}
\arguments{
\item{x}{an \R object.}

\item{\dots}{further arguments passed to or from methods.}
}
\description{
Generic print.nice function
}
\section{Methods (by class)}{
\itemize{
\item \code{nice.default}: simply calls \code{\link{print}}
}}

