\name{print.nice}
\alias{print.nice}
\alias{print.nice.default}
\title{Generic print.nice Function and Default Method}

\usage{
print.nice(x, ...)

<<<<<<< HEAD
\S3method{print.nice}{default}(x, ...)
=======
\method{print.nice}{default}(x, ...)
>>>>>>> 1c1c3802ecd2c79b4b03494fe38c6139a3e6db36
}

\arguments{
\item{x}{an \R object.}
\item{\dots}{further arguments passed to or from methods.}
}

\description{
Generic print.nice function and default method.
}


